moving profile related methods to new class
